refractor(settings): break up the settings into one module per configuration
- moves urls.py to project_root
- moves wsgi.py to project_root
- 'config.settings' is replaced by 'config', selectively applicable
  configs are imported in '__init__.py'

fixes: #47
