+++ {{cookiecutter.repo_name}}/Procfile
@@ -1,1 +1,1 @@
-web: gunicorn --pythonpath="$PWD/{{cookiecutter.repo_name}}" config.wsgi:application
+web: gunicorn --pythonpath="$PWD/{{cookiecutter.repo_name}}" wsgi:application
