+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/config/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import
+
+from .local import Local  # noqa
+from .production import Production  # noqa
