+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/manage.py
@@ -3,7 +3,7 @@ import os
 import sys
 
 if __name__ == "__main__":
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config")
     os.environ.setdefault("DJANGO_CONFIGURATION", "Local")
 
     from configurations.management import execute_from_command_line
