+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/urls.py
