+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/wsgi.py
@@ -19,7 +19,7 @@ import os
 # if running multiple sites in the same mod_wsgi process. To fix this, use
 # mod_wsgi daemon mode with each site in its own daemon process, or use
 # os.environ["DJANGO_SETTINGS_MODULE"] = "{{ repo_name }}.settings"
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config")
 os.environ.setdefault("DJANGO_CONFIGURATION", "Production")
 
 # This application object is used by any WSGI server configured to use this
